+++ app/frontend/services/fyApi/models/FleetVehicleExport.ts
@@ -1,0 +1,27 @@
+/* generated using openapi-typescript-codegen -- do no edit */
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export type FleetVehicleExport = {
+    name: string;
+    slug: string;
+    shipCode?: string;
+    manufacturerName: string;
+    manufacturerCode?: string;
+    paintSlug?: string;
+    shipName?: string;
+    shipSerial?: string;
+    wanted: boolean;
+    flagship: boolean;
+    public: boolean;
+    saleNotify: boolean;
+    nameVisible: boolean;
+    groups: Array<string>;
+    modules: Array<string>;
+    upgrades: Array<string>;
+    username?: string | null;
+    userAvatar?: string | null;
+    userRsiHandle?: string | null;
+};
+