+++ app/frontend/services/fyApi/models/VehiclePublic.ts
@@ -18,8 +18,9 @@ export type VehiclePublic = {
     hangarGroups: Array<HangarGroupPublic>;
     loaner: boolean;
     model: Model;
-    username?: string;
-    userAvatar?: string;
+    username?: string | null;
+    userAvatar?: string | null;
+    userRsiHandle?: string | null;
     modelModuleIds: Array<string>;
     modelUpgradeIds: Array<string>;
     modulePackage?: ModelModulePackage;